feat: useGpt 훅 생성
